fix: persist snackbars in case of an error
